requirements.txt
absl-py==2.2.2 ; python_version >= "3.13"
annotated-types==0.7.0 ; python_version >= "3.13"
anyio==4.9.0 ; python_version >= "3.13"
authlib==1.5.2 ; python_version >= "3.13"
cachetools==5.5.2 ; python_version >= "3.13"
certifi==2025.4.26 ; python_version >= "3.13"
cffi==1.17.1 ; python_version >= "3.13" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.2 ; python_version >= "3.13"
click==8.2.1 ; python_version >= "3.13"
cloudpickle==3.1.1 ; python_version >= "3.13"
colorama==0.4.6 ; python_version >= "3.13" and platform_system == "Windows"
cryptography==45.0.2 ; python_version >= "3.13"
deprecated==1.2.18 ; python_version >= "3.13"
docstring-parser==0.13 ; python_version >= "3.13"
fastapi==0.115.12 ; python_version >= "3.13"
google-adk==0.1.0 ; python_version >= "3.13"
google-api-core==2.24.2 ; python_version >= "3.13"
google-api-python-client==2.169.0 ; python_version >= "3.13"
google-auth-httplib2==0.2.0 ; python_version >= "3.13"
google-auth==2.40.2 ; python_version >= "3.13"
google-cloud-aiplatform==1.94.0 ; python_version >= "3.13"
google-cloud-bigquery==3.33.0 ; python_version >= "3.13"
google-cloud-core==2.4.3 ; python_version >= "3.13"
google-cloud-resource-manager==1.14.2 ; python_version >= "3.13"
google-cloud-secret-manager==2.23.3 ; python_version >= "3.13"
google-cloud-speech==2.32.0 ; python_version >= "3.13"
google-cloud-storage==2.19.0 ; python_version >= "3.13"
google-cloud-trace==1.16.1 ; python_version >= "3.13"
google-crc32c==1.7.1 ; python_version >= "3.13"
google-genai==1.16.1 ; python_version >= "3.13"
google-resumable-media==2.7.2 ; python_version >= "3.13"
googleapis-common-protos==1.70.0 ; python_version >= "3.13"
graphviz==0.20.3 ; python_version >= "3.13"
greenlet==3.2.2 ; python_version == "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
grpc-google-iam-v1==0.14.2 ; python_version >= "3.13"
grpcio-status==1.71.0 ; python_version >= "3.13"
grpcio==1.71.0 ; python_version >= "3.13"
h11==0.16.0 ; python_version >= "3.13"
httpcore==1.0.9 ; python_version >= "3.13"
httplib2==0.22.0 ; python_version >= "3.13"
httpx-sse==0.4.0 ; python_version >= "3.13"
httpx==0.28.1 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
importlib-metadata==8.6.1 ; python_version >= "3.13"
mcp==1.9.1 ; python_version >= "3.13"
numpy==2.2.6 ; python_version >= "3.13"
opentelemetry-api==1.33.1 ; python_version >= "3.13"
opentelemetry-exporter-gcp-trace==1.9.0 ; python_version >= "3.13"
opentelemetry-resourcedetector-gcp==1.9.0a0 ; python_version >= "3.13"
opentelemetry-sdk==1.33.1 ; python_version >= "3.13"
opentelemetry-semantic-conventions==0.54b1 ; python_version >= "3.13"
packaging==25.0 ; python_version >= "3.13"
proto-plus==1.26.1 ; python_version >= "3.13"
protobuf==5.29.4 ; python_version >= "3.13"
pyasn1-modules==0.4.2 ; python_version >= "3.13"
pyasn1==0.6.1 ; python_version >= "3.13"
pycparser==2.22 ; python_version >= "3.13" and platform_python_implementation != "PyPy"
pydantic-core==2.33.2 ; python_version >= "3.13"
pydantic-settings==2.9.1 ; python_version >= "3.13"
pydantic==2.11.4 ; python_version >= "3.13"
pyparsing==3.2.3 ; python_version >= "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.13"
python-dotenv==1.1.0 ; python_version >= "3.13"
python-multipart==0.0.20 ; python_version >= "3.13"
pyyaml==6.0.2 ; python_version >= "3.13"
requests==2.32.3 ; python_version >= "3.13"
rsa==4.2 ; python_version >= "3.13"
shapely==2.1.1 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
sniffio==1.3.1 ; python_version >= "3.13"
sqlalchemy==2.0.41 ; python_version >= "3.13"
sse-starlette==2.3.5 ; python_version >= "3.13"
starlette==0.46.2 ; python_version >= "3.13"
typing-extensions==4.13.2 ; python_version >= "3.13"
typing-inspection==0.4.1 ; python_version >= "3.13"
tzdata==2025.2 ; python_version >= "3.13" and platform_system == "Windows"
tzlocal==5.3.1 ; python_version >= "3.13"
uritemplate==4.1.1 ; python_version >= "3.13"
urllib3==2.4.0 ; python_version >= "3.13"
uvicorn==0.34.2 ; python_version >= "3.13"
websockets==15.0.1 ; python_version >= "3.13"
wrapt==1.17.2 ; python_version >= "3.13"
zipp==3.21.0 ; python_version >= "3.13"
